// src/model/rest/report/transport/ReportTurnoverCapitalsourceTransport.ts
export type ReportTurnoverCapitalsourceTransport = {
  capitalsourceType: number;
  capitalsourceState: number;
  capitalsourceComment: string;
  amountBeginOfMonthFixed: number;
  amountEndOfMonthFixed: number;
  amountEndOfMonthCalculated: number;
  amountCurrent: number;
  amountCurrentState: string;
};